Turtle Pong
Remake of Pong using the Turtle module in Python. There are some slight changes, such as the ball being replaced with a turtle, and the speed of the ball randomly changing to make the game more challenging. It is a two player game that utilizes the up and down arrow keys, as well as the W and S keys for player controls.